+++ packages/ubcli/lib/dbScripts/oracle.js
@@ -12,24 +12,24 @@ const fs = require('fs')
  * @param {Object} databaseConfig A database configuration
  */
 module.exports.dropDatabase = function dropDatabase (session, databaseConfig) {
-  let conn = session.connection
-  let upperUser = databaseConfig.userID.toUpperCase()
+  const conn = session.connection
+  const upperUser = databaseConfig.userID.toUpperCase()
 
-  let userExist = conn.xhr({endpoint: 'runSQL', URLParams: {CONNECTION: DBA_FAKE}, data: "SELECT COUNT(1) as CNT FROM dba_users WHERE username = '" + upperUser + "'"})
+  const userExist = conn.xhr({ endpoint: 'runSQL', URLParams: { CONNECTION: DBA_FAKE }, data: "SELECT COUNT(1) as CNT FROM dba_users WHERE username = '" + upperUser + "'" })
   if (userExist.length && userExist[0].CNT !== 0) {
-    let activeConnections = conn.xhr({
+    const activeConnections = conn.xhr({
       endpoint: 'runSQL',
-      URLParams: {CONNECTION: DBA_FAKE},
+      URLParams: { CONNECTION: DBA_FAKE },
       data: "SELECT sid, serial# AS sn FROM v$session WHERE username = '" + upperUser + "'"
     })
     for (let i = 0, l = activeConnections.length; i < l; i++) {
       conn.xhr({
         endpoint: 'runSQL',
-        URLParams: {CONNECTION: DBA_FAKE},
+        URLParams: { CONNECTION: DBA_FAKE },
         data: "alter system kill session '" + activeConnections[i].SID + ', ' + activeConnections[i].SN + "'"
       })
     }
-    conn.xhr({endpoint: 'runSQL', URLParams: {CONNECTION: DBA_FAKE}, data: 'DROP USER ' + upperUser + ' CASCADE'})
+    conn.xhr({ endpoint: 'runSQL', URLParams: { CONNECTION: DBA_FAKE }, data: 'DROP USER ' + upperUser + ' CASCADE' })
   } else {
     console.warn('User %s dose not exists. Drop skipped', upperUser)
   }
@@ -43,11 +43,11 @@ module.exports.dropDatabase = function dropDatabase (session, databaseConfig) {
 module.exports.createDatabase = function createDatabase (conn, databaseConfig) {
   conn.xhr({
     endpoint: 'runSQL',
-    URLParams: {CONNECTION: DBA_FAKE},
+    URLParams: { CONNECTION: DBA_FAKE },
     data: `CREATE USER ${databaseConfig.userID} IDENTIFIED BY ${databaseConfig.password} DEFAULT TABLESPACE USERS TEMPORARY TABLESPACE TEMP PROFILE DEFAULT ACCOUNT UNLOCK;`
   })
 
-  let grants = [
+  const grants = [
     'GRANT RESOURCE, CONNECT, CTXAPP TO {0}',
     'ALTER USER {0} DEFAULT ROLE ALL',
     'GRANT ALTER ANY INDEX TO {0}',
@@ -78,7 +78,7 @@ module.exports.createDatabase = function createDatabase (conn, databaseConfig) {
   for (let i = 0, l = grants.length; i < l; i++) {
     conn.xhr({
       endpoint: 'runSQL',
-      URLParams: {CONNECTION: DBA_FAKE},
+      URLParams: { CONNECTION: DBA_FAKE },
       data: grants[i].replace('{0}', databaseConfig.userID)
     })
   }
@@ -93,30 +93,30 @@ module.exports.createDatabase = function createDatabase (conn, databaseConfig) {
 module.exports.createMinSchema = function createMinSchema (conn, clientNum, databaseConfig) {
   conn.xhr({
     endpoint: 'runSQL',
-    URLParams: {CONNECTION: databaseConfig.name},
+    URLParams: { CONNECTION: databaseConfig.name },
     data: `CREATE SEQUENCE SEQ_UBMAIN START WITH ${clientNum}0000000000 MAXVALUE ${clientNum}4999999999 MINVALUE ${clientNum}0000000000 NOCYCLE CACHE 10 ORDER`
   })
 
   conn.xhr({
     endpoint: 'runSQL',
-    URLParams: {CONNECTION: databaseConfig.name},
+    URLParams: { CONNECTION: databaseConfig.name },
     data: `CREATE SEQUENCE SEQ_UBMAIN_BY1 START WITH ${clientNum}500000000000 MAXVALUE ${clientNum}999999999999 MINVALUE ${clientNum}500000000000 NOCYCLE ORDER`
   })
 
-  let createObjectSQL = fs.readFileSync(path.join(__dirname, 'oracleObjects.sql'), 'utf8')
-  let delimRe = /\r\n/.test(createObjectSQL) ? '/\r\n--' : '/\n--' // git can remove \r\n
+  const createObjectSQL = fs.readFileSync(path.join(__dirname, 'oracleObjects.sql'), 'utf8')
+  const delimRe = /\r\n/.test(createObjectSQL) ? '/\r\n--' : '/\n--' // git can remove \r\n
   let statements = createObjectSQL.split(delimRe)
   statements.forEach(function (statement) {
     if (statement) {
-      conn.xhr({endpoint: 'runSQL', URLParams: {CONNECTION: databaseConfig.name}, data: statement})
+      conn.xhr({ endpoint: 'runSQL', URLParams: { CONNECTION: databaseConfig.name }, data: statement })
     }
   })
 
-  let initialData = fs.readFileSync(path.join(__dirname, 'oracleTables.sql'), 'utf8')
+  const initialData = fs.readFileSync(path.join(__dirname, 'oracleTables.sql'), 'utf8')
   statements = initialData.split(delimRe)
   statements.forEach(function (statement) {
     if (statement) {
-      conn.xhr({endpoint: 'runSQL', URLParams: {CONNECTION: 'main'}, data: statement})
+      conn.xhr({ endpoint: 'runSQL', URLParams: { CONNECTION: 'main' }, data: statement })
     }
   })
 }