+++ packages/ubcli/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - detailed explanation of database creation in initDB - see https://unitybase.info/api/server-v5/module-initDB.html
+ for tips how to create a database manually
 
 ### Changed